+++ src/utility/style_utility.tsx
@@ -36,6 +36,9 @@ export function getEmphasisClasses(emphasis: EmphasisType, color: ColorOptionTyp
     } else if (emphasis === 'medium') {
 	
         c = ` background-color--${color}-light border--px-0 ${getTextColorClass()}`;
+        c += ` ${getBorderClass()}`
+        c += ` border--solid`
+        c += ` border--${color}-light`
         if (options.hasRipple) {
             c += ` ripple ripple--${color}-normal`
         }